Prototype method to load a movie resource using libav.
This is a proof of concept method to load a movie resource using libav. It
provides IO adaptor functions for MHKArchive and shows that by lying about the
size of the file and by doing an initial seek to the movie's offset in the
archive, libav can properly parse the mov structure.